test-requirements.txt
pytest
pytest-trio
pre-commit
black==20.8b1
